feat: 레스토랑 목록 정렬 및 필터링 로직 추가
